+++ noodles-vcf/src/header/record/value/map/contig.rs
@@ -11,7 +11,9 @@ use std::fmt;
 use super::{Fields, Indexed, Inner, Map, TryFromFieldsError};
 
 type StandardTag = tag::Standard;
-type Tag = super::tag::Tag<StandardTag>;
+
+/// A VCF header contig map tag.
+pub type Tag = super::tag::Tag<StandardTag>;
 
 /// An inner VCF header contig map value.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]